warning fixes, expanded lims, made ints to longs for data capabilities

+++ src/uptime_pop.c
@@ -1,7 +1,7 @@
 #include "includes/pop_module.h"
 
 // Global tick counter
-static unsigned int tick_counter = 0;
+static unsigned long long tick_counter = 0;
 
 void int_to_str(int value, char* buffer) {
     char temp[10];
